requirements.txt
annotated-types
anyio
certifi
cffi
charset-normalizer
click
cryptography
docker
fastapi
h11
idna
pycparser
pydantic
pydantic_core
requests
sniffio
starlette
typing_extensions
urllib3
uvicorn
jinja2
python-multipart
pyyaml
